basereader: Variables does get_variables_*
